PoC_Step4/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pandas
openpyxl
python-docx
pypdf
pdfplumber
